+++ lib/src/logic/yandex_gpt_api.dart
@@ -1,20 +1,19 @@
 import 'package:dio/dio.dart';
-import 'package:yandex_gpt_rest_api/src/logic/api/yandex_gpt_api.dart';
 import 'package:yandex_gpt_rest_api/src/logic/client/yandex_gpt_header_interceptor.dart';
 import 'package:yandex_gpt_rest_api/src/logic/client/yandex_gpt_http_client.dart';
 import 'package:yandex_gpt_rest_api/src/models/models.dart';
 import 'package:yandex_gpt_rest_api/src/utils/constants/api_url.dart';
 
-final class YandexGptApiClient implements YandexGptApi {
+final class YandexGptApi {
   final YandexGptHttpClient _client;
   final YandexGptHeaderInterceptor _headerInterceptor;
 
-  YandexGptApiClient({
+  YandexGptApi({
     required AuthToken token,
     String? catalog,
   }) : this.withDio(dio: Dio(), token: token, catalog: catalog);
 
-  YandexGptApiClient.withOptions({
+  YandexGptApi.withOptions({
     required BaseOptions options,
     required AuthToken token,
     String? catalog,
@@ -24,7 +23,7 @@ final class YandexGptApiClient implements YandexGptApi {
           catalog: catalog,
         );
 
-  YandexGptApiClient.withDio({
+  YandexGptApi.withDio({
     required Dio dio,
     required AuthToken token,
     String? catalog,
@@ -36,12 +35,10 @@ final class YandexGptApiClient implements YandexGptApi {
     dio.interceptors.add(_headerInterceptor);
   }
 
-  @override
   void changeToken(AuthToken token) {
     _headerInterceptor.changeToken(token);
   }
 
-  @override
   Future<TextGenerationResponse> generateText(
     TextGenerationRequest request, {
     CancelToken? cancelToken,
@@ -56,7 +53,6 @@ final class YandexGptApiClient implements YandexGptApi {
     );
   }
 
-  @override
   Future<TextGenerationAsyncResponse> generateAsyncText(
     TextGenerationRequest request, {
     CancelToken? cancelToken,
@@ -69,7 +65,6 @@ final class YandexGptApiClient implements YandexGptApi {
     return TextGenerationAsyncResponse.fromJson(res);
   }
 
-  @override
   Future<TextGenerationAsyncResponse> getOperationTextGenerate(
     String operationId, {
     CancelToken? cancelToken,
@@ -81,7 +76,6 @@ final class YandexGptApiClient implements YandexGptApi {
     return TextGenerationAsyncResponse.fromJson(res);
   }
 
-  @override
   Future<EmbeddingResponse> getTextEmbedding(
     EmbeddingRequest request, {
     CancelToken? cancelToken,
@@ -94,7 +88,6 @@ final class YandexGptApiClient implements YandexGptApi {
     return EmbeddingResponse.fromJson(res);
   }
 
-  @override
   Future<TokenizeResponse> tokenizeCompletion(
     TextGenerationRequest request, {
     CancelToken? cancelToken,
@@ -107,7 +100,6 @@ final class YandexGptApiClient implements YandexGptApi {
     return TokenizeResponse.fromJson(res);
   }
 
-  @override
   Future<TokenizeResponse> tokenizeText(
     TokenizeTextRequest request, {
     CancelToken? cancelToken,