+++ lib/yandex_gpt_rest_api.dart
@@ -1,3 +1,2 @@
-export 'src/logic/api/yandex_gpt_api.dart';
-export 'src/logic/client/yandex_gpt_api_client.dart';
+export 'src/logic/yandex_gpt_api.dart';
 export 'src/models/models.dart';